Uncomplete work history section if edited
