+++ app/controllers/candidate_interface/work_history/destroy_controller.rb
@@ -13,6 +13,8 @@ module CandidateInterface
         .find(work_experience_params[:id])
         .destroy!
 
+      current_application.update!(work_history_completed: false)
+
       redirect_to candidate_interface_work_history_show_path
     end
