+++ app/controllers/candidate_interface/work_history/edit_controller.rb
@@ -30,6 +30,7 @@ module CandidateInterface
         else
           redirect_to candidate_interface_work_history_show_path
         end
+        current_application.update!(work_history_completed: false)
       else
         render :new
       end
@@ -49,6 +50,8 @@ module CandidateInterface
       @work_experience_form = WorkExperienceForm.new(work_experience_form_params)
 
       if @work_experience_form.update(work_experience)
+        current_application.update!(work_history_completed: false)
+
         redirect_to candidate_interface_work_history_show_path
       else
         render :edit
